* commit 'e31fe2b20d41698e59ef0e02e5aa286a176517a7': (741 commits)
  update SUNDIALS version to 3.0.0
  Updated documentation figures (PDF, PNG and EPS) from sundials.figs.pptx file
  add note to install guide on building shared Sundials libs
  Updated vector module list for all figures, solver and matrix list for all figures and example directories for all packages.
  CVODES manual in place.  Minor cleanup of IDA and IDAS manuals for things along the way.  Corrected subsectioning in the linsol chapter.
  remove check on precision for hypre
  remove mpi compiler from cvode
  fix undefined symbols in shared libs
  First draft of adjoint part fo manuel.  IDAS now has a complete draft.  It still has some overfull hboxes to deal with, but labels are all resolved.
  Remove external dependencies from all Sundials solvers
  remove pdf of partial generic linsol text
  minor change to nvec doc revisions
  NVector documentation udpates
  adjust table placement and spacing
  First draft of IDAS manuel except for eth idas_adj.tex file.  Minor changes to teh IDA manuel to make these as synchronized as possible.
  SP: More implicit type conversion fixes in CUDA module.
  SP: Fixed some indentation the automatic tool missed.
  SP: Correct data type for grid size in for CUDA thread partitioning.
  SP: Correct indentation in Vector classes for RAJA and CUDA modules.
  add sunlinsol version variables
  ...

+++ src/nvec_pargrid/fnvector_parallel_grid.h
@@ -60,8 +60,8 @@ extern N_Vector F2C_KINSOL_vec;
 
 extern N_Vector F2C_ARKODE_vec;
 
-/* 
- * Prototypes of exported functions 
+/*
+ * Prototypes of exported functions
  *
  * FNV_INITPG    - initializes parallel grid vector operations for main problem
  * FNV_INITPG_Q  - initializes parallel grid vector operations for quadratures
@@ -75,19 +75,19 @@ extern N_Vector F2C_ARKODE_vec;
 #define MPI_Fint int
 #endif
 
-void FNV_INITPG(MPI_Fint *comm, int *code, long int *dims, long int *dim_len, 
-		long int *dim_alen, long int *dim_off, long int *F_ordering, 
+void FNV_INITPG(MPI_Fint *comm, int *code, int *dims, long int *dim_len,
+		long int *dim_alen, long int *dim_off, int *F_ordering,
 		long int *glob_len, int *ier);
-void FNV_INITPG_Q(MPI_Fint *comm, int *code, long int *dims, long int *dim_lenQ, 
-		  long int *dim_alenQ, long int *dim_offQ, long int *F_ordering, 
+void FNV_INITPG_Q(MPI_Fint *comm, int *code, int *dims, long int *dim_lenQ,
+		  long int *dim_alenQ, long int *dim_offQ, int *F_ordering,
 		  long int *glob_lenQ, int *ier);
-void FNV_INITPG_B(MPI_Fint *comm, int *code, long int *dims, long int *dim_lenB, 
-		  long int *dim_alenB, long int *dim_offB, long int *F_ordering, 
+void FNV_INITPG_B(MPI_Fint *comm, int *code, int *dims, long int *dim_lenB,
+		  long int *dim_alenB, long int *dim_offB, int *F_ordering,
 		  long int *glob_lenB, int *ier);
-void FNV_INITPG_QB(MPI_Fint *comm, int *code, long int *dims, long int *dim_lenQB, 
-		   long int *dim_alenQB, long int *dim_offQB, long int *F_ordering, 
+void FNV_INITPG_QB(MPI_Fint *comm, int *code, int *dims, long int *dim_lenQB,
+		   long int *dim_alenQB, long int *dim_offQB, int *F_ordering,
 		   long int *glob_lenQB, int *ier);
-void FNV_INITPG_S(int *code, long int *Ns, int *ier);
+void FNV_INITPG_S(int *code, int *Ns, int *ier);
 
 #ifdef __cplusplus
 }
